requirements.txt
arcade==2.6.14; python_version >= "3.7"
attrs==21.4.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
cffi==1.15.0; python_version >= "3.7"
pillow==9.1.1; python_version >= "3.7"
pycparser==2.21; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
pyglet==2.0.dev16; python_version >= "3.7"
pymunk==6.2.1; python_version >= "3.7"
pytiled-parser==2.0.1; python_version >= "3.7"
typing-extensions==4.2.0; python_version >= "3.7"
